+++ changelog.md
@@ -7,6 +7,7 @@ All notable changes to this mod will be documented in this file.
 - Added sound effects to ability switching on and off
   - SFX shortened from testing feedback
   - Sound folder moved to the proper location
+- Hullmod changed to "Integrated Forge" and all references updated
 
 ## [0.1.21] - 2026-05-06